Move add-edit custom fields to own component
